% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customised_model_yardstick.R
\name{customised_model_yardstick}
\alias{customised_model_yardstick}
\title{customised_model_yardstick}
\usage{
customised_model_yardstick()
}
\value{
a list with model parameters built on top of use_kknn(sex ~ ., data = penguin_train)
}
\description{
customised_model_yardstick
}
\examples{
if(interactive()){
personalysed_knn_penguins_sex<-customised_model_yardstick()
}
}
