% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer.R
\name{simulator}
\alias{simulator}
\title{simulator}
\usage{
simulator(numer = 25, bins = 50)
}
\arguments{
\item{numer}{integer length of distribution}

\item{bins}{integer size of bins for histogram}
}
\value{
a histigram showing CI at the 95% confidence level
}
\description{
simulator
}
\examples{
simulator()
}
